refactor: ready for adding other chat apps
